# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Novel Adaptations API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Novel Adaptations",
      "url": "https://noveladaptations.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://noveladaptations.com",
      "description": "Production (apex; www.noveladaptations.com serves the same API)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "na_session",
        "description": "Session cookie set by POST /api/v1/auth/verify (HttpOnly, 30 days)."
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Not signed in. `{ \"error\": { \"code\": \"unauthorized\", \"message\": \"\u2026\" } }`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "unauthorized",
                "message": "Sign in to continue."
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Signed in, but not an admin.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "forbidden",
                "message": "Admin access required."
              }
            }
          }
        }
      },
...
      "Unprocessable": {
        "description": "Validation failure \u2014 always 422 in this API, never 400.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "validation_error",
                "message": "Enter a valid email address."
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected server failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "example": {
              "error": {
                "code": "internal_error",
                "message": "Something went wrong."
              }
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "not_found"
              },
              "message": {
                "type": "string",
                "example": "Adaptation not found."
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/admin/backfill/tmdb": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Run one TMDB enrichment batch",
        "description": "Enriches up to `n` poster-less screen works via TMDB. `n` must be an\ninteger 1\u201310; default 5. Failures are counted per-row, not fatal.\n",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "n",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10,
              "default": 5
            },
            "description": "Batch size. Default 5."
          }
        ],
        "responses": {
          "200": {
            "description": "Batch complete.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "done",
                    "enriched",
                    "failed",
                    "remaining"
                  ],
                  "properties": {
                    "done": {
                      "type": "integer",
                      "description": "Rows attempted.",
                      "example": 5
                    },
                    "enriched": {
                      "type": "integer",
                      "description": "Rows with a TMDB hit.",
                      "example": 4
                    },
                    "failed": {
                      "type": "integer",
                      "description": "Rows with no TMDB match, no key, or an exception.",
                      "example": 1
                    },
                    "remaining": {
                      "type": "integer",
                      "description": "Rows still needing enrichment AFTER this batch.",
                      "example": 61
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}